move type extensions and types to root of src directory
