+++ src/type_extensions.ts
@@ -1,0 +1,15 @@
+import type {
+  StorageLayoutDiffConfig,
+  StorageLayoutDiffUserConfig,
+} from './types';
+import 'hardhat/types/config';
+
+declare module 'hardhat/types/config' {
+  interface HardhatConfig {
+    storageLayoutDiff: StorageLayoutDiffConfig;
+  }
+
+  interface HardhatUserConfig {
+    storageLayoutDiff?: StorageLayoutDiffUserConfig;
+  }
+}
